fix bit operating bug

+++ client-tez/src/main/java/org/apache/tez/runtime/library/common/sort/buffer/WriteBufferManager.java
@@ -427,7 +427,7 @@ public class WriteBufferManager<K, V> {
     return waitSendBuffers;
   }
 
-  private int getNextSeqNo(int partitionId) {
+  private long getNextSeqNo(int partitionId) {
     partitionToSeqNo.putIfAbsent(partitionId, 0);
     int seqNo = partitionToSeqNo.get(partitionId);
     partitionToSeqNo.put(partitionId, seqNo + 1);
